🔥 upgrade decoder class
