+++ base_decoder.rb
@@ -1,0 +1,14 @@
+require "./nameable"
+
+class BaseDecorator < Nameable
+  attr_accessor :nameable
+
+  def initialize(nameable)
+    super(nameable.name)
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
